Adding MainView class
