+++ bin/.gitignore
@@ -1,4 +1,1 @@
 /com/
-/gui$1.class
-/gui$2.class
-/gui.class
